+++ test/golden/hasql-features/Queries.hs
@@ -7,9 +7,9 @@ where
 
 import Queries.Internal
 import Queries.Types
-import Queries.FindUserByName as Queries
-import Queries.FindUsers as Queries
-import Queries.FindPosts as Queries
-import Queries.FindMembers as Queries
-import Queries.GetEvent as Queries
-import Queries.DeleteUsers as Queries
+import Queries.FindUserByName as Queries hiding (paramsEncoder, rowDecoder)
+import Queries.FindUsers as Queries hiding (paramsEncoder, rowDecoder)
+import Queries.FindPosts as Queries hiding (paramsEncoder, rowDecoder)
+import Queries.FindMembers as Queries hiding (paramsEncoder, rowDecoder)
+import Queries.GetEvent as Queries hiding (paramsEncoder, rowDecoder)
+import Queries.DeleteUsers as Queries hiding (paramsEncoder, rowDecoder)