+++ test/golden/hasql-features/Queries/DeleteUsers.hs
@@ -8,9 +8,18 @@
 {-# LANGUAGE TypeFamilies #-}
 module Queries.DeleteUsers where
 
-import Queries.Internal (Query(..), Enum, Params, Result, ToRow(..), FromRow(..), ToField(..), FromField(..))
+import Queries.Internal (Query(..), Enum, Params)
+-- The row type is named `Result`, and so is IsStatement's associated type. GHC
+-- rejects a qualified name on the left of an associated type instance, so it is
+-- the row family that gets qualified here instead.
+import qualified Queries.Internal
+import qualified Data.Int
+import qualified Data.Vector
 import qualified Hasql.Decoders
 import qualified Hasql.Encoders
+import qualified Hasql.Mapping.IsScalar
+import qualified Hasql.Mapping.IsStatement
+import qualified Hasql.Statement
 
 import qualified Data.Foldable
 import qualified Data.Functor.Contravariant
@@ -22,19 +31,26 @@ data instance Params "DeleteUsers" = Params_DeleteUsers
   {
   }
 
-data instance Result "DeleteUsers" = Result_DeleteUsers
+data instance Queries.Internal.Result "DeleteUsers" = Result_DeleteUsers
   {
   }
 
-instance ToRow (Params "DeleteUsers") where
-  {-# INLINE toRow #-}
-  toRow =
-    mconcat
-      [       ]
+paramsEncoder :: Hasql.Encoders.Params (Params "DeleteUsers")
+paramsEncoder =
+  mconcat
+    [     ]
+{-# INLINE paramsEncoder #-}
 
-instance FromRow (Result "DeleteUsers") where
-  {-# INLINE fromRow #-}
-  fromRow =
-    pure Result_DeleteUsers
+rowDecoder :: Hasql.Decoders.Row (Queries.Internal.Result "DeleteUsers")
+rowDecoder =
+  pure Result_DeleteUsers
+{-# INLINE rowDecoder #-}
+
+instance Hasql.Mapping.IsStatement.IsStatement (Params "DeleteUsers") where
+  type Result (Params "DeleteUsers") = ()
+  statement =
+    Hasql.Statement.preparable sql paramsEncoder (Hasql.Decoders.noResult)
+    where
+      Query sql = query_DeleteUsers