+++ test/golden/hasql-features/Queries/FindMembers.hs
@@ -8,9 +8,18 @@
 {-# LANGUAGE TypeFamilies #-}
 module Queries.FindMembers where
 
-import Queries.Internal (Query(..), Enum, Params, Result, ToRow(..), FromRow(..), ToField(..), FromField(..))
+import Queries.Internal (Query(..), Enum, Params)
+-- The row type is named `Result`, and so is IsStatement's associated type. GHC
+-- rejects a qualified name on the left of an associated type instance, so it is
+-- the row family that gets qualified here instead.
+import qualified Queries.Internal
+import qualified Data.Int
+import qualified Data.Vector
 import qualified Hasql.Decoders
 import qualified Hasql.Encoders
+import qualified Hasql.Mapping.IsScalar
+import qualified Hasql.Mapping.IsStatement
+import qualified Hasql.Statement
 
 import qualified Queries.Types
 import qualified GHC.Base
@@ -25,25 +34,32 @@ data instance Params "FindMembers" = Params_FindMembers
     role :: (Queries.Types.Enum "organization_role")
   }
 
-data instance Result "FindMembers" = Result_FindMembers
+data instance Queries.Internal.Result "FindMembers" = Result_FindMembers
   {
     role :: !((Queries.Types.Enum "organization_role")),
     previous_role :: !(GHC.Base.Maybe ((Queries.Types.Enum "organization_role")))
   }
 
-instance ToRow (Params "FindMembers") where
-  {-# INLINE toRow #-}
-  toRow =
-    mconcat
-      [ 
-      Data.Functor.Contravariant.contramap (\Params_FindMembers{..} -> role) (Hasql.Encoders.param (Hasql.Encoders.nonNullable toField))
-      ]
+paramsEncoder :: Hasql.Encoders.Params (Params "FindMembers")
+paramsEncoder =
+  mconcat
+    [ 
+    Data.Functor.Contravariant.contramap (\Params_FindMembers{..} -> role) (Hasql.Encoders.param (Hasql.Encoders.nonNullable Hasql.Mapping.IsScalar.encoder))
+    ]
+{-# INLINE paramsEncoder #-}
 
-instance FromRow (Result "FindMembers") where
-  {-# INLINE fromRow #-}
-  fromRow =
-    pure Result_FindMembers
-      <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable fromField)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nullable fromField)
+rowDecoder :: Hasql.Decoders.Row (Queries.Internal.Result "FindMembers")
+rowDecoder =
+  pure Result_FindMembers
+    <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Mapping.IsScalar.decoder)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nullable Hasql.Mapping.IsScalar.decoder)
+{-# INLINE rowDecoder #-}
+
+instance Hasql.Mapping.IsStatement.IsStatement (Params "FindMembers") where
+  type Result (Params "FindMembers") = Data.Vector.Vector (Queries.Internal.Result "FindMembers")
+  statement =
+    Hasql.Statement.preparable sql paramsEncoder (Hasql.Decoders.rowVector rowDecoder)
+    where
+      Query sql = query_FindMembers