+++ test/golden/simple-query-hasql/Queries/ListUsers.hs
@@ -8,9 +8,18 @@
 {-# LANGUAGE TypeFamilies #-}
 module Queries.ListUsers where
 
-import Queries.Internal (Query(..), Enum, Params, Result, ToRow(..), FromRow(..), ToField(..), FromField(..))
+import Queries.Internal (Query(..), Enum, Params)
+-- The row type is named `Result`, and so is IsStatement's associated type. GHC
+-- rejects a qualified name on the left of an associated type instance, so it is
+-- the row family that gets qualified here instead.
+import qualified Queries.Internal
+import qualified Data.Int
+import qualified Data.Vector
 import qualified Hasql.Decoders
 import qualified Hasql.Encoders
+import qualified Hasql.Mapping.IsScalar
+import qualified Hasql.Mapping.IsStatement
+import qualified Hasql.Statement
 
 import qualified Data.Int
 import qualified Data.Text
@@ -30,7 +39,7 @@ data instance Params "ListUsers" = Params_ListUsers
     age :: Data.Int.Int32
   }
 
-data instance Result "ListUsers" = Result_ListUsers
+data instance Queries.Internal.Result "ListUsers" = Result_ListUsers
   {
     id :: !(Data.Int.Int32),
     name :: !(Data.Text.Text),
@@ -43,26 +52,33 @@ data instance Result "ListUsers" = Result_ListUsers
     avatar :: !(GHC.Base.Maybe Data.ByteString.ByteString)
   }
 
-instance ToRow (Params "ListUsers") where
-  {-# INLINE toRow #-}
-  toRow =
-    mconcat
-      [ 
-      Data.Functor.Contravariant.contramap (\Params_ListUsers{..} -> age) (Hasql.Encoders.param (Hasql.Encoders.nonNullable Hasql.Encoders.int4))
-      ]
+paramsEncoder :: Hasql.Encoders.Params (Params "ListUsers")
+paramsEncoder =
+  mconcat
+    [ 
+    Data.Functor.Contravariant.contramap (\Params_ListUsers{..} -> age) (Hasql.Encoders.param (Hasql.Encoders.nonNullable Hasql.Encoders.int4))
+    ]
+{-# INLINE paramsEncoder #-}
 
-instance FromRow (Result "ListUsers") where
-  {-# INLINE fromRow #-}
-  fromRow =
-    pure Result_ListUsers
-      <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.int4)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.varchar)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nullable Hasql.Decoders.text)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.bpchar)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.bool)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.numeric)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nullable Hasql.Decoders.float8)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.jsonb)
-      <*> Hasql.Decoders.column (Hasql.Decoders.nullable Hasql.Decoders.bytea)
+rowDecoder :: Hasql.Decoders.Row (Queries.Internal.Result "ListUsers")
+rowDecoder =
+  pure Result_ListUsers
+    <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.int4)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.varchar)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nullable Hasql.Decoders.text)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.bpchar)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.bool)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.numeric)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nullable Hasql.Decoders.float8)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nonNullable Hasql.Decoders.jsonb)
+    <*> Hasql.Decoders.column (Hasql.Decoders.nullable Hasql.Decoders.bytea)
+{-# INLINE rowDecoder #-}
+
+instance Hasql.Mapping.IsStatement.IsStatement (Params "ListUsers") where
+  type Result (Params "ListUsers") = Data.Vector.Vector (Queries.Internal.Result "ListUsers")
+  statement =
+    Hasql.Statement.preparable sql paramsEncoder (Hasql.Decoders.rowVector rowDecoder)
+    where
+      Query sql = query_ListUsers